Rename auth transaction classes
