+++ Auth0.podspec
@@ -6,13 +6,15 @@ web_auth_files = [
   'Auth0/ObjectiveC/A0SHA.h',
   'Auth0/ObjectiveC/A0SHA.m',
   'Auth0/Array+Encode.swift',
+  'Auth0/ASCallbackTransaction.swift',
+  'Auth0/ASTransaction.swift',
   'Auth0/AuthCancelable.swift',
   'Auth0/AuthProvider.swift',
   'Auth0/AuthSession.swift',
   'Auth0/AuthTransaction.swift',
-  'Auth0/AuthenticationServicesSession.swift',
-  'Auth0/AuthenticationServicesSessionCallback.swift',
   'Auth0/Auth0WebAuth.swift',
+  'Auth0/BaseCallbackTransaction.swift',
+  'Auth0/BaseTransaction.swift',
   'Auth0/BioAuthentication.swift',
   'Auth0/ClaimValidators.swift',
   'Auth0/IDTokenSignatureValidator.swift',
@@ -25,8 +27,6 @@ web_auth_files = [
   'Auth0/NSURLComponents+OAuth2.swift',
   'Auth0/OAuth2Grant.swift',
   'Auth0/ResponseType.swift',
-  'Auth0/SessionCallbackTransaction.swift',
-  'Auth0/SessionTransaction.swift',
   'Auth0/TransactionStore.swift',
   'Auth0/WebAuth.swift',
   'Auth0/WebAuthError.swift',
